present ForgotPin without react navigation
