+++ packages/core-mobile/app/screens/mainView/ForgotPin.tsx
@@ -1,22 +1,19 @@
-import React, { useCallback } from 'react'
-import { useNavigation } from '@react-navigation/native'
+import React from 'react'
 import AvaText from 'components/AvaText'
 import { Button, View } from '@avalabs/k2-mobile'
 import { Space } from 'components/Space'
 import FlexSpacer from 'components/FlexSpacer'
 import SlideToConfirm from 'screens/earn/components/SlideToConfirm'
 
-const ForgotPin = ({ onConfirm }: { onConfirm: () => void }): JSX.Element => {
-  const { goBack, canGoBack } = useNavigation()
-
-  const onClose = useCallback(() => {
-    if (canGoBack()) {
-      goBack()
-    }
-  }, [canGoBack, goBack])
-
+const ForgotPin = ({
+  onConfirm,
+  onCancel
+}: {
+  onConfirm: () => void
+  onCancel: () => void
+}): JSX.Element => {
   return (
-    <View style={{ flex: 1, padding: 16 }}>
+    <View style={{ padding: 16, height: '100%' }}>
       <AvaText.LargeTitleBold>
         {`Do you want to\nreset your PIN?`}
       </AvaText.LargeTitleBold>
@@ -32,7 +29,7 @@ const ForgotPin = ({ onConfirm }: { onConfirm: () => void }): JSX.Element => {
         type="tertiary"
         size="xlarge"
         style={{ marginTop: 16 }}
-        onPress={onClose}>
+        onPress={onCancel}>
         Cancel
       </Button>
     </View>
